Move GenerateUrl to LinkCommandBase
Move GenerateUrl from LocalRepositoryModel to LinkCommandBase.
