+++ src/GitHub.VisualStudio/Commands/BlameLinkCommand.cs
@@ -15,8 +15,8 @@ namespace GitHub.VisualStudio.Commands
     public class BlameLinkCommand : LinkCommandBase, IBlameLinkCommand
     {
         [ImportingConstructor]
-        protected BlameLinkCommand(IGitHubServiceProvider serviceProvider)
-            : base(CommandSet, CommandId, serviceProvider)
+        protected BlameLinkCommand(IGitHubServiceProvider serviceProvider, Lazy<IGitService> gitService)
+            : base(CommandSet, CommandId, serviceProvider, gitService)
         {
         }
