+++ src/GitHub.VisualStudio/Commands/LinkCommandBase.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Globalization;
 using System.IO;
 using System.Threading.Tasks;
 using GitHub.Api;
@@ -20,16 +21,19 @@ namespace GitHub.VisualStudio.Commands
         static readonly ILogger log = LogManager.ForContext<LinkCommandBase>();
         readonly Lazy<ISimpleApiClientFactory> apiFactory;
         readonly Lazy<IUsageTracker> usageTracker;
+        readonly Lazy<IGitService> lazyGitService;
 
         protected LinkCommandBase(
             Guid commandSet,
             int commandId,
-            IGitHubServiceProvider serviceProvider)
+            IGitHubServiceProvider serviceProvider,
+            Lazy<IGitService> gitService)
             : base(commandSet, commandId)
         {
             ServiceProvider = serviceProvider;
             apiFactory = new Lazy<ISimpleApiClientFactory>(() => ServiceProvider.TryGetService<ISimpleApiClientFactory>());
             usageTracker = new Lazy<IUsageTracker>(() => serviceProvider.TryGetService<IUsageTracker>());
+            lazyGitService = gitService;
         }
 
         protected ILocalRepositoryModel ActiveRepo { get; private set; }
@@ -134,7 +138,7 @@ namespace GitHub.VisualStudio.Commands
 
             var repo = GetRepositoryByPath(activeDocument.Name);
 
-            return repo.GenerateUrl(linkType, activeDocument.Name, activeDocument.StartLine, activeDocument.EndLine);
+            return GenerateUrl(lazyGitService.Value, repo, linkType, activeDocument.Name, activeDocument.StartLine, activeDocument.EndLine);
         }
 
         protected override void QueryStatus()
@@ -159,6 +163,98 @@ namespace GitHub.VisualStudio.Commands
                 dir = dir.Parent;
             }
             return false;
+        }
+
+        /// <summary>
+        /// Generates a http(s) url to the repository in the remote server, optionally
+        /// pointing to a specific file and specific line range in it.
+        /// </summary>
+        /// <param name="linkType">Type of link to generate</param>
+        /// <param name="path">The file to generate an url to. Optional.</param>
+        /// <param name="startLine">A specific line, or (if specifying the <paramref name="endLine"/> as well) the start of a range</param>
+        /// <param name="endLine">The end of a line range on the specified file.</param>
+        /// <returns>An UriString with the generated url, or null if the repository has no remote server configured or if it can't be found locally</returns>
+        public static async Task<UriString> GenerateUrl(IGitService gitService,
+            ILocalRepositoryModel repo, LinkType linkType, string path = null, int startLine = -1, int endLine = -1)
+        {
+            if (repo.CloneUrl == null)
+                return null;
+
+            var sha = await gitService.GetLatestPushedSha(path ?? repo.LocalPath);
+            // this also incidentally checks whether the repo has a valid LocalPath
+            if (String.IsNullOrEmpty(sha))
+                return repo.CloneUrl.ToRepositoryUrl().AbsoluteUri;
+
+            if (path != null && Path.IsPathRooted(path))
+            {
+                // if the path root doesn't match the repository local path, then ignore it
+                if (!path.StartsWith(repo.LocalPath, StringComparison.OrdinalIgnoreCase))
+                {
+                    System.Diagnostics.Debug.Assert(false, String.Format(CultureInfo.CurrentCulture, "GenerateUrl: path {0} doesn't match repository {1}", path, repo.LocalPath));
+                    path = null;
+                }
+                else
+                    path = path.Substring(repo.LocalPath.Length + 1);
+            }
+
+            if (startLine > 0 && endLine > 0 && startLine > endLine)
+            {
+                // if startLine is greater than endLine and both are set, swap them
+                var temp = startLine;
+                startLine = endLine;
+                endLine = temp;
+            }
+
+            if (startLine == endLine)
+            {
+                // if startLine is the same as endLine don't generate a range link
+                endLine = -1;
+            }
+
+            return new UriString(GenerateUrl(linkType, repo.CloneUrl.ToRepositoryUrl().AbsoluteUri, sha, path, startLine, endLine));
+        }
+
+        const string CommitFormat = "{0}/commit/{1}";
+        const string BlobFormat = "{0}/blob/{1}/{2}";
+        const string BlameFormat = "{0}/blame/{1}/{2}";
+        const string StartLineFormat = "{0}#L{1}";
+        const string EndLineFormat = "{0}-L{1}";
+        static string GenerateUrl(LinkType linkType, string basePath, string sha, string path, int startLine = -1, int endLine = -1)
+        {
+            if (sha == null)
+                return basePath;
+
+            if (String.IsNullOrEmpty(path))
+                return String.Format(CultureInfo.InvariantCulture, CommitFormat, basePath, sha);
+
+            var ret = String.Format(CultureInfo.InvariantCulture, GetLinkFormat(linkType), basePath, sha, path.Replace(@"\", "/"));
+
+            if (startLine < 0)
+                return ret;
+            ret = String.Format(CultureInfo.InvariantCulture, StartLineFormat, ret, startLine);
+            if (endLine < 0)
+                return ret;
+            return String.Format(CultureInfo.InvariantCulture, EndLineFormat, ret, endLine);
+        }
+
+        /// <summary>
+        /// Selects the proper format for the link type, defaults to the blob url when link type is not selected.
+        /// </summary>
+        /// <param name="linkType">Type of link to generate</param>
+        /// <returns>The string format of the selected link type</returns>
+        static string GetLinkFormat(LinkType linkType)
+        {
+            switch (linkType)
+            {
+                case LinkType.Blame:
+                    return BlameFormat;
+
+                case LinkType.Blob:
+                    return BlobFormat;
+
+                default:
+                    return BlobFormat;
+            }
         }
     }
 }
