+++ test/GitHub.VisualStudio.UnitTests/Commands/LinkCommandBaseTests.cs
@@ -1,14 +1,13 @@
-﻿using System.Threading.Tasks;
-using System.Collections.Generic;
+﻿using System;
+using System.Threading.Tasks;
 using GitHub.Models;
 using GitHub.Exports;
 using GitHub.Services;
-using GitHub.Primitives;
+using GitHub.VisualStudio.Commands;
 using NSubstitute;
-using LibGit2Sharp;
 using NUnit.Framework;
 
-public class LocalRepositoryModelTests : TestBaseClass
+public class LinkCommandBaseTests : TestBaseClass
 {
     [TestCase(1, LinkType.Blob, false, "https://github.com/foo/bar", "123123", @"src\dir\file1.cs", -1, -1, "https://github.com/foo/bar/blob/123123/src/dir/file1.cs")]
     [TestCase(2, LinkType.Blob, false, "https://github.com/foo/bar", "123123", @"src\dir\file1.cs", 1, -1, "https://github.com/foo/bar/blob/123123/src/dir/file1.cs#L1")]
@@ -48,32 +47,18 @@ public class LocalRepositoryModelTests : TestBaseClass
             if (createRootedPath && path != null)
                 path = System.IO.Path.Combine(basePath.FullName, path);
             ILocalRepositoryModel model = null;
-            if (!string.IsNullOrEmpty(baseUrl))
-                model = new LocalRepositoryModel("bar", new UriString(baseUrl), basePath.FullName, gitService);
-            else
-                model = new LocalRepositoryModel(basePath.FullName, gitService);
-            var result = await model.GenerateUrl(linkType, path, startLine, endLine);
-            Assert.That(expected, Is.EqualTo(result?.ToString()));
+            model = new LocalRepositoryModel { Name = "bar", CloneUrl = baseUrl, LocalPath = basePath.FullName };
+
+            var result = await LinkCommandBase.GenerateUrl(gitService, model, linkType, path, startLine, endLine);
+
+            Assert.That(result?.ToString(), Is.EqualTo(expected));
         }
     }
 
     static IGitService CreateGitService(string sha)
     {
         var gitservice = Substitute.For<IGitService>();
-        var repo = Substitute.For<IRepository>();
-        gitservice.GetRepository(Args.String).Returns(repo);
         gitservice.GetLatestPushedSha(Args.String).Returns(Task.FromResult(sha));
-        if (!string.IsNullOrEmpty(sha))
-        {
-            var refs = Substitute.For<ReferenceCollection>();
-            var refrence = Substitute.For<Reference>();
-            refs.ReachableFrom(Arg.Any<IEnumerable<Reference>>(), Arg.Any<IEnumerable<Commit>>()).Returns(new Reference[] { refrence });
-            repo.Refs.Returns(refs);
-            var commit = Substitute.For<Commit>();
-            commit.Sha.Returns(sha);
-            repo.Commits.Returns(new FakeCommitLog { commit });
-        }
-
         return gitservice;
     }
 }
